refactor: make invited page as dialog
